feat: move resolvers to resolver files and db components into the appropriate files in the db folder, and fix imports
