+++ app/db/db_management.py
@@ -1,0 +1,5 @@
+from pymongo import MongoClient
+
+client = MongoClient("mongodb://localhost:27017/")
+db = client["users_db"]
+users_collection = db["users"]
