+++ app/db/db_models.py
@@ -5,3 +5,4 @@ from pydantic import BaseModel
 class DbUser(BaseModel):
     userId: str
     favorites: List[str]
+
